+++ Assets/LDtkUnity/CHANGELOG.md
@@ -35,6 +35,7 @@ This update focuses on significant performance optimisation and a revamp to sepa
 ## Fixes
 - Importing backup files for separate level files are now ignored in addition to project files
 - Fixed the ILDtkImportedLevel not invoking when no level fields were defined
+- Fixed the level background position being incorrect when set to "Not scaled" or "Fit inside"
 - Updated samples
 
 ### Breaking Changes